Add files via upload
---Kill the terrorists!---
This is a game where the player is a brave warrior who has to kill the terrorists who attacked India!
All the instructions are given on the home page of the game. 
Press 's' to start the game.
Use right and left arrow keys to move.
Press 'space' to shoot.
Shoot 50 terrorists to protect India and its residents from terrorists.
Good luck warriors!